requirements.txt
asttokens==2.0.5
backcall==0.2.0
black==21.12b0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.10
click==8.0.3
cycler==0.11.0
decorator==5.1.1
descartes==1.1.0
distlib==0.3.4
executing==0.8.2
filelock==3.4.2
flake8==4.0.1
fonttools==4.29.0
identify==2.4.5
idna==3.3
ipython==8.0.1
jedi==0.18.1
kiwisolver==1.3.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mizani==0.7.3
mypy==0.931
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.22.1
packaging==21.3
palettable==3.3.0
pandas==1.4.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
plotnine==0.8.0
pre-commit==2.17.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
scipy==1.7.3
-e git+ssh://git@github.com/mbmackenzie/setlist-survival.git@50807c00cbd507d89d08b562405718e8a50d6349#egg=setlist_survival
six==1.16.0
stack-data==0.1.4
statsmodels==0.13.1
tidybear==0.0.3
toml==0.10.2
tomli==1.2.3
traitlets==5.1.1
types-PyYAML==6.0.3
typing_extensions==4.0.1
urllib3==1.26.8
virtualenv==20.13.0
wcwidth==0.2.5
